+++ temporian/core/test/registered_operators_test.py
@@ -98,6 +98,7 @@ class RegisteredOperatorsTest(absltest.TestCase):
             "TIMESTAMPS",
             "UNIQUE_TIMESTAMPS",
             "UNTIL_NEXT",
+            "WHERE",
             "XOR",
         ]