+++ temporian/implementation/numpy/test/registered_operators_test.py
@@ -96,6 +96,7 @@ class RegisteredOperatorsTest(absltest.TestCase):
             "TIMESTAMPS",
             "UNIQUE_TIMESTAMPS",
             "UNTIL_NEXT",
+            "WHERE",
             "XOR",
         ]